move more things to defcfg.rs
